requirements.txt
piexif==1.1.3
pillow==11.1.0
PyExifTool==0.5.6